Create a SILType wrapper for TypeBase::isSuperclassOf.
TypeBase::isSuperclassOf recursively determines if one type is a superclass of
another. I need this functionality to handle specialization of
unconditional_checked_cast archetype_to_archetype. Specifically if the
specialized types are sub/super classes I need to be able to ascertain that so I
can insert an upcast/downcast appropriately.

It makes sense to add this method to SILType itself since we should avoid
touching the AST as much as possible at the SIL level in order to make SIL as
self contained as we can.

rdar://16490450

Swift SVN r15746