Optimize calls to the builtin canBeObjCClass when possible.
Attempt to simplify calls to canBeObjCClass to literals when we know the
type. We already do this in silgen, but more opportunities to simplify
can happen as the result of devirtualization and inlining.

Eliminates 46 of 50 static calls to canBeObjCClass in the stdlib, and 7
of 11 in the RC4 benchmark.

Fixes <rdar://problem/16543013>.

Swift SVN r16044