Only examine box container pointer when determining whether to promote.
Implement JoeG's suggestion to limit the examination of uses to the
container pointer since there is no legal way for the address pointer to
outlive the box container.

Deletes 16 of the 45 alloc_box in the stdlib in my build
configuration.

canValueEscape() can be simplified now since it is only being used
examine how partial_apply get used, but I'll leave that for a separate
commit.

Swift SVN r16047