Suppress duplicated initializers that come from importing class factories as initializers.
As noted in the FIXME, this is a hack. We need to do a bit of
infrastructure work before we can detect and suppress duplicates like
this in a meaningful way. Name lookup already does a similar thing for
ambiguity-resolution purposes. However, doing this here makes the
printed modules much cleaner.

look significantly more reasonable.


Swift SVN r16531