Now that we have labeled breaks as a fallback, implement the first part of: <rdar://problem/16563853> 'break' should be able to break out of switch statements
We start by just *rejecting* unlabeled breaks that are in switch statements, which
are currently used to break out to the nearest loop.  These will experience a behavior
change, and rejecting them for awhile will let people who are using this notice.

Fun fact: Apparently breaking through a switch was never used in the testsuite or 
stdlib.


Swift SVN r16617