[please update llvm] Debug info: Use the mangled name as unique identifier for types where possible.
We use the DICompositeType's UniqueIdentifier field to store the mangled
name, the name field to store the base name and Apple-llvm will emit the
unique identifier as DW_AT_name if RunTimeLang == Swift. This way the
metadata format is compatible between swift and non-Apple clang.
As bonus, we get mangled-name-based type uniquing for swift.

<rdar://problem/16303510> apple-types accelerator table should include basename for swift types

Swift SVN r16665