fix <rdar://problem/16660680> QoI: fatal() in init method complains about super.init being called multiple times
The problem here is that the implicitly generated self.init() call is unreachable, 
and DCE runs after DI, so DI doesn't know it is unreachable.  Fix this by adding a trivial
reachability check before emitting a diagnostic. 

This isn't a wholesome solution though, as certain infinite loops and other scenarios that
rely on constant folding won't be handled properly.  If anyone runs into this, we could consider
moving constant prop and DCE earlier, though that comes with its own tradeoffs.


Swift SVN r16898