Don't create branches with arguments to blocks that do not expect them.
I am planning on doing a larger change to ensure that blocks
consistently have a BB arg if they are targeted by a conditional branch
that has a payload (e.g. a switch_enum) so that we do not have to
special case like this.

Fix <rdar://problem/16761996>.

Swift SVN r17238