Grab BB args from the correct block.
We do not necessarily want to grab the BB arg that we will use in the
branch we are creating from the same block that the switch_enum we are
eliminating is in.

There is still a latent bug here for the case where the block we are
creating a branch to expects an argument but the block we would grab the
argument from doesn't have an argument. This will be fixed when I land
changes that enforce that payload producing ops like switch_enum target
branches that always have a BB arg.

Swift SVN r17239